ref : WorkLog 도메인 삭제 및 Settlement 도메인으로 리팩터링
