+++ homecare/src/main/java/jaega/homecare/domain/serviceMatch/repository/DashboardStats.java
@@ -1,0 +1,7 @@
+package jaega.homecare.domain.serviceMatch.repository;
+
+public record DashboardStats(
+        Long totalCaregivers,
+        Long assignedCaregivers,
+        Long waitingApplicants
+) {}
