+++ homecare/src/main/java/jaega/homecare/domain/serviceMatch/repository/ServiceMatchQueryRepository.java
@@ -1,19 +1,28 @@
 package jaega.homecare.domain.serviceMatch.repository;
 
+import com.querydsl.core.Tuple;
+import com.querydsl.core.types.ExpressionUtils;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
+import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jaega.homecare.domain.caregiver.entity.QCaregiver;
+import jaega.homecare.domain.caregiverCenter.entity.CaregiverStatus;
+import jaega.homecare.domain.caregiverCenter.entity.QCaregiverCenter;
 import jaega.homecare.domain.consumer.entity.QConsumer;
+import jaega.homecare.domain.serviceMatch.entity.MatchStatus;
 import jaega.homecare.domain.serviceMatch.entity.QServiceMatch;
 import jaega.homecare.domain.serviceRequest.entity.QServiceRequest;
+import jaega.homecare.domain.settlement.dto.res.GetDashboardWorkStatusResponse;
+import jaega.homecare.domain.settlement.dto.res.WorkPlaceDistribution;
 import jaega.homecare.domain.users.entity.QUser;
 import jaega.homecare.domain.users.entity.ServiceType;
-import jaega.homecare.domain.workLog.dto.res.GetCaregiverMatchesResponse;
+import jaega.homecare.domain.settlement.dto.res.GetCaregiverMatchesResponse;
 import jaega.homecare.domain.caregiver.repository.CaregiverRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
+import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -125,6 +134,92 @@ public class ServiceMatchQueryRepository {
                         base.hourlyWage(),
                         base.status(),
                         base.notes()
+                ))
+                .toList();
+    }
+
+    /**
+     *  대시보드
+     */
+
+
+    public DashboardStats getDashboardStats(UUID centerId, LocalDate date) {
+        QCaregiver caregiver = QCaregiver.caregiver;
+        QServiceMatch serviceMatch = QServiceMatch.serviceMatch;
+        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
+
+        List<Tuple> result = queryFactory
+                .select(
+                        caregiver.countDistinct(), // 센터의 오늘 매칭이 배정, 완료된 요양보호사 수 조회
+                        ExpressionUtils.as(
+                                JPAExpressions
+                                        .select(serviceMatch.caregiver.countDistinct())
+                                        .from(serviceMatch)
+                                        .join(serviceMatch.caregiver, caregiver)
+                                        .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
+                                        .where(
+                                                caregiverCenter.center.centerId.eq(centerId)
+                                                        .and(serviceMatch.serviceDate.eq(date))
+                                                        .and(serviceMatch.matchStatus.in(MatchStatus.PENDING, MatchStatus.COMPLETED))
+                                        ), "assignedCaregivers"
+                        ),
+                        ExpressionUtils.as( // 매칭이 있으나 오늘이 아닌 요양보호사 수 조회
+                                JPAExpressions
+                                        .select(serviceMatch.count())
+                                        .from(serviceMatch)
+                                        .where(serviceMatch.matchStatus.eq(MatchStatus.PENDING)
+                                                .and(serviceMatch.serviceDate.eq(date))), "waitingApplicants"
+                        )
+                )
+                .from(caregiver)
+                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
+                .where(caregiverCenter.center.centerId.eq(centerId))
+                .fetch();
+
+        Tuple row = result.get(0);
+        return new DashboardStats(row.get(0, Long.class), row.get(1, Long.class), row.get(2, Long.class));
+    }
+
+    // 요양 보호사 대시보드의 근무지 별 분포 통계 조회
+    public List<WorkPlaceDistribution> getWorkPlaceDistributionByServiceType(UUID centerId) {
+        QCaregiver caregiver = QCaregiver.caregiver;
+        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
+
+        // 센터 소속 활동 중인 요양보호사(+서비스타입) 조회
+        List<Tuple> rows = queryFactory
+                .select(caregiver, caregiver.serviceTypes)
+                .from(caregiver)
+                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
+                .where(
+                        caregiverCenter.center.centerId.eq(centerId),
+                        caregiverCenter.status.eq(CaregiverStatus.ACTIVE)
+                )
+                .fetch();
+
+        if (rows.isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        // ServiceType별 카운트 계산
+        Map<ServiceType, Long> serviceTypeCount = new HashMap<>();
+        for (Tuple row : rows) {
+            Set<ServiceType> serviceTypes = row.get(caregiver.serviceTypes);
+            if (serviceTypes != null) {
+                for (ServiceType st : serviceTypes) {
+                    serviceTypeCount.put(st, serviceTypeCount.getOrDefault(st, 0L) + 1);
+                }
+            }
+        }
+
+        // 총합 계산
+        long total = serviceTypeCount.values().stream().mapToLong(Long::longValue).sum();
+
+        // DTO 변환
+        return serviceTypeCount.entrySet().stream()
+                .map(entry -> new WorkPlaceDistribution(
+                        entry.getKey(),
+                        entry.getValue(),
+                        total > 0 ? (double) entry.getValue() * 100 / total : 0.0
                 ))
                 .toList();
     }
