+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/req/CreateSettlementRequest.java
@@ -1,0 +1,13 @@
+package jaega.homecare.domain.settlement.dto.req;
+
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.UUID;
+
+public record CreateSettlementRequest(
+        UUID caregiverCenterId,
+        UUID serviceMatchId,
+        Double distanceLog
+) {
+}
