+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/req/GetWorkLogRequest.java
@@ -1,0 +1,8 @@
+package jaega.homecare.domain.settlement.dto.req;
+
+import java.util.UUID;
+
+public record GetWorkLogRequest (
+        UUID workMatchId
+) {
+}
