+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/res/GetCaregiverSettlementSummaryResponse.java
@@ -1,0 +1,11 @@
+package jaega.homecare.domain.settlement.dto.res;
+
+import java.math.BigDecimal;
+
+public record GetCaregiverSettlementSummaryResponse(
+        BigDecimal totalAmount,
+        Long completedCount,
+        Long plannedCount,
+        Long cancelledCount
+) {
+}
