+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/res/GetDashboardSettlementResponse.java
@@ -1,0 +1,10 @@
+package jaega.homecare.domain.settlement.dto.res;
+
+import java.math.BigDecimal;
+
+public record GetDashboardSettlementResponse(
+        BigDecimal totalSettledAmount,
+        Long unsettledCount,
+        Long fraudAlertCount  // 부정행위 알림 건수
+) {
+}
