+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/res/GetMonthlyPaymentResponse.java
@@ -1,0 +1,10 @@
+package jaega.homecare.domain.settlement.dto.res;
+
+import java.math.BigDecimal;
+
+public record GetMonthlyPaymentResponse(
+        int year,
+        int month,
+        BigDecimal totalAmount
+) {
+}
