+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/res/GetSettlementByPaid.java
@@ -1,0 +1,11 @@
+package jaega.homecare.domain.settlement.dto.res;
+
+import java.time.LocalDate;
+import java.util.UUID;
+
+public record GetSettlementByPaid(
+        UUID settlementId,
+        LocalDate serviceDate,
+        String caregiverName
+) {
+}
