+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/res/GetSettlementCenterSummaryResponse.java
@@ -1,0 +1,10 @@
+package jaega.homecare.domain.settlement.dto.res;
+
+import java.math.BigDecimal;
+
+public record GetSettlementCenterSummaryResponse(
+        BigDecimal totalAmount,
+        Long completedCount,
+        Long pendingCount,
+        Long rejectedCount
+) {}
