+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/res/GetSettlementResponse.java
@@ -1,0 +1,13 @@
+package jaega.homecare.domain.settlement.dto.res;
+
+import java.time.LocalTime;
+import java.util.UUID;
+
+public record GetSettlementResponse(
+        UUID settlementId,
+        LocalTime serviceStartTime,
+        LocalTime serviceEndTime,
+        Double distanceLog,
+        Boolean isPaid
+) {
+}
