+++ homecare/src/main/java/jaega/homecare/domain/settlement/dto/res/WorkPlaceDistribution.java
@@ -1,0 +1,9 @@
+package jaega.homecare.domain.settlement.dto.res;
+
+import jaega.homecare.domain.users.entity.ServiceType;
+
+public record WorkPlaceDistribution (
+        ServiceType serviceType,
+        Long count,
+        Double percent
+) {}
