--- homecare/src/main/java/jaega/homecare/domain/workLog/dto/req/CreateWorkLogRequest.java
@@ -1,16 +1,0 @@
-package jaega.homecare.domain.workLog.dto.req;
-
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.UUID;
-
-public record CreateWorkLogRequest(
-        UUID caregiverId,
-        LocalTime workStartTime,
-        LocalTime workEndTime,
-        LocalDate workDate,
-        String workAddress,
-        Double distanceLog
-) {
-}
