--- homecare/src/main/java/jaega/homecare/domain/workLog/dto/res/GetCaregiverSettlementSummaryResponse.java
@@ -1,11 +1,0 @@
-package jaega.homecare.domain.workLog.dto.res;
-
-import java.math.BigDecimal;
-
-public record GetCaregiverSettlementSummaryResponse(
-        BigDecimal totalAmount,
-        Long completedCount,
-        Long plannedCount,
-        Long cancelledCount
-) {
-}
