--- homecare/src/main/java/jaega/homecare/domain/workLog/dto/res/GetSettlementCenterSummaryResponse.java
@@ -1,10 +1,0 @@
-package jaega.homecare.domain.workLog.dto.res;
-
-import java.math.BigDecimal;
-
-public record GetSettlementCenterSummaryResponse(
-        BigDecimal totalAmount,
-        Long completedCount,
-        Long pendingCount,
-        Long rejectedCount
-) {}
